requirements.txt
customtkinter==5.2.1
darkdetect==0.8.0
packaging==23.2
pyFirmata==1.1.0
pyserial==3.5
